- progress in creating landing page

+++ Spring/src/main/java/App/controllers/CarController.java
@@ -3,7 +3,7 @@ package App.controllers;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,7 +13,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import App.DTO.CarDTO;
 import App.entities.Car;
 import App.entities.User;
 import App.repositories.CarRepository;
@@ -41,34 +40,29 @@ public class CarController {
     }
 
     @PostMapping
-    public ResponseEntity<String> addCar(@RequestBody Car car) {
-        try {
-            // Optional<User> userOptional = this.userRepository.findById();
-            // //throw new Exception(car.getUser().toString());
-
-            // // Optional<User> userOptional = this.userRepository.findById(car.getUser().getId());
-
-            // if (!userOptional.isPresent()) {
-            //     throw new Exception("there is no user with this id");
-            // }
-
-            // car.setUser(userOptional.get());
-            this.carRepository.save(car);
-            return ResponseEntity.ok("Data was succefully created"); 
-        } catch (DataIntegrityViolationException ex) {
-            return ResponseEntity.badRequest().body(ex.getMessage());
-        } catch (Exception ex) {
-            return ResponseEntity.badRequest().body(ex.getMessage());
+    public ResponseEntity<?> createCar(@RequestBody Car car) {
+        User user = car.getUser();
+        if (user == null || !userRepository.existsById(user.getId())) {
+            return new ResponseEntity<>("There is no user with this id", HttpStatus.NOT_FOUND);
         }
+
+        carRepository.save(car);
+        return new ResponseEntity<>("Car was added to database", HttpStatus.CREATED);
     }
 
     @DeleteMapping("{carId}")
-    public void deleteCar(@PathVariable("carId") Integer id) {
+    public ResponseEntity<?> deleteCar(@PathVariable("carId") Integer id) {
+        if (!this.carRepository.existsById(id)) {
+            return new ResponseEntity<>("There is no car with this Id", HttpStatus.NOT_FOUND);
+        }
+
         this.carRepository.deleteById(id);
+        return new ResponseEntity<>("Car was succesfully deleted", HttpStatus.OK);
     }
 
     @DeleteMapping
-    public void deleteAllCars() {
+    public ResponseEntity<?> deleteAllCars() {
         this.carRepository.deleteAll();
+        return new ResponseEntity<>("All cars were succesfully deleted", HttpStatus.OK);
     }
 }
