+++ Spring/src/main/java/App/controllers/UserController.java
@@ -3,6 +3,8 @@ package App.controllers;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -36,13 +38,22 @@ public class UserController {
     }
 
     @PostMapping
-    public void addUser(@RequestBody User user) {
+    public ResponseEntity<?> addUser(@RequestBody User user) {
+        if (this.userRepository.findByEmail(user.getEmail()) != null) {
+            return new ResponseEntity<>("There is already user with this email", HttpStatus.CONFLICT);
+        }
+
         this.userRepository.save(user);
+        return new ResponseEntity<>("User was successfully added", HttpStatus.CREATED);
     }
 
     @DeleteMapping("{userId}")
-    public void deleteUser(@PathVariable("userId") Integer id) {
+    public ResponseEntity<?> deleteUser(@PathVariable("userId") Integer id) {
+        if (this.userRepository.findById(id) == null) {
+            return new ResponseEntity<>("There is no user with this id", HttpStatus.NOT_FOUND);
+        }
         this.userRepository.deleteById(id);
+        return new ResponseEntity<>("User was successfully deleted", HttpStatus.OK);
     }
 
     @DeleteMapping
